% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_probelmatic_variables.r
\name{clean_problematic_variables}
\alias{clean_problematic_variables}
\title{Clean and Prepare Data for Analysis}
\usage{
clean_problematic_variables(dataframe)
}
\arguments{
\item{dataframe}{A data.frame or data.table intended for analysis.}
}
\value{
A data.table where logical and character columns have been converted to factors, and all data is in a consistent format for analysis.
}
\description{
This function converts the input data frame to a data.table (if not already), checks for unsupported data types (complex or raw),
and converts logical or character columns to factors. It stops with a comprehensive error message if unsupported data types are found
or if "AsIs" objects with logical or character mode are detected. It aims to prepare data for analysis by ensuring that it is in a consistent format.
}
\examples{
\dontrun{
  # Assuming `data.table` is already loaded
  my_data <- data.table(a = 1:10, b = rnorm(10), c = LETTERS[1:10], d = I(matrix(1:4, 2)))
  clean_data <- cleanAndPrepareData(my_data)
  print(clean_data)
}
}
