% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_info.R
\name{column_info}
\alias{column_info}
\title{Column Information}
\usage{
column_info(dt, corr_variable = NULL)
}
\arguments{
\item{dt}{A data.table or data frame.}

\item{corr_variable}{An optional string specifying a column name to calculate correlations with.}
}
\value{
A data.table with information on each column.
}
\description{
This function returns a data.table with information about each column in a provided data table.
}
\examples{
dt <- data.table(x = 1:10, y = rnorm(10))
column_info(dt)
}
