% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_columns_to_numeric.r
\name{convert_columns_to_numeric}
\alias{convert_columns_to_numeric}
\title{Convert Character Columns to Numeric Where Applicable}
\usage{
convert_columns_to_numeric(dt, tolerance = 0.95)
}
\arguments{
\item{dt}{A data.table or data frame whose character columns are to be evaluated for conversion to numeric.}

\item{tolerance}{A numeric value between 0 and 1 indicating the minimum proportion of elements in a character
column that need to be convertible to numeric for the column to be converted. Defaults to 0.95.}
}
\value{
The original data.table or data frame with applicable character columns converted to numeric.
}
\description{
Iterates over the columns of a data.table or data frame, converting those character columns to numeric
which are predominantly numeric according to a specified tolerance level. Columns that are not character
or do not meet the numeric content threshold are left unchanged.
}
\examples{
\dontrun{
  library(data.table)
  dt <- data.table(a = c("1", "2", "3"), b = c("a", "b", "c"), c = c("4.5", "5.6", "NA"))
  dt <- convert_columns_to_numeric(dt)
  # Columns 'a' and 'c' are converted to numeric, while 'b' remains as character
}
}
