% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_data.r
\name{create_dummy_data}
\alias{create_dummy_data}
\title{Create Dummy Data from Original Data with Configurable Noise Percentage}
\usage{
create_dummy_data(data, percent = 10)
}
\arguments{
\item{data}{A data.table or data frame containing the original data.}

\item{percent}{The maximum percentage of the original value by which to adjust.
Default is 10 (for 10\%).}
}
\value{
A data.table with dummy data for numeric columns.
}
\description{
This function iterates over each numeric column of a given data.table or data frame,
adding noise expressed as a percentage of each value to generate dummy data. The process
maintains the original sign of non-zero numeric values and ensures that zeros and NAs remain
unchanged. Non-numeric columns are left as-is, without any alteration. Please make sure
that they do not contain sensitive information. If the dataset includes sensitive
information in non-numeric columns, it is the user's responsibility to anonymize or remove
these columns. The aim is to facilitate the creation of a non-sensitive, dummy dataset that
retains the statistical properties of the original data for analysis or sharing purposes,
without compromising the confidentiality of the original information.
}
\examples{
\dontrun{
  # Assuming `data.table` is already loaded
  original_data <- data.table(a = 1:10, b = rnorm(10), c = LETTERS[1:10])
  dummy_data <- create_dummy_data(original_data, percent = 5)
  print(dummy_data)
}
}
