% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveMeDataTable.r
\name{moveMeDataTable}
\alias{moveMeDataTable}
\title{Move Columns in a Data Table}
\usage{
moveMeDataTable(data, tomove, where = "first", ba = NULL)
}
\arguments{
\item{data}{A data.table whose columns are to be rearranged.}

\item{tomove}{A vector of column names or indices to move. This can be a mix of numeric indices and column names.}

\item{where}{A character string indicating where to move the specified columns.
Valid options are "first", "last", "before", and "after". The default is "first".}

\item{ba}{A column name (as a character string) indicating the base column relative to which the columns
specified in \code{tomove} should be moved. This parameter is required if \code{where} is "before" or "after".}
}
\value{
A data.table with the columns rearranged according to the specified criteria.
}
\description{
This function rearranges the columns of a data.table based on specified criteria.
It allows columns to be moved to the first or last position, or before or after a specific column.
}
\details{
This function modifies the column order of a data.table without altering the data within.
When using "before" or "after", ensure the \code{ba} parameter is correctly specified, or an error will be thrown.
}
\examples{
\dontrun{
  library(data.table)
  dt <- data.table(A = 1:5, B = letters[1:5], C = rnorm(5))
  # Move column B to the first position
  moveMeDataTable(dt, tomove = "B", where = "first")
  # Move columns B and C to the last position
  moveMeDataTable(dt, tomove = c("B", "C"), where = "last")
  # Move column A before column C
  moveMeDataTable(dt, tomove = "A", where = "before", ba = "C")
  # Move column A after column B
  moveMeDataTable(dt, tomove = "A", where = "after", ba = "B")
}
}
