% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_from_path.r
\name{read_files_from_path}
\alias{read_files_from_path}
\title{Read Multiple Data Files from a Directory into a Named List}
\arguments{
\item{path}{Character. The path to the folder containing the files. Backslashes (\verb{\\\\}) are supported.}

\item{extensions}{Character vector. A vector of file extensions (without dots) to read, such as \code{c("xlsx", "csv")}.}
}
\value{
A named list of data frames or tibbles, one for each file. The names correspond to the original filenames (without extensions).
}
\description{
This function reads all files in a specified directory that match the provided file extensions
and loads them into R as a named list. Each list element is named after the file (without extension),
and the file is read using the appropriate method based on its extension.
}
\details{
Supported file extensions and their readers:
\itemize{
\item \code{csv} - Read using \code{read.csv()}
\item \code{xlsx}, \code{xls} - Read using \code{readxl::read_excel()}
}
Files with unsupported extensions will result in an error.
}
\examples{
\dontrun{
data_list <- read_files_from_path("W:/PROJECTS/MESN-WSER_2025/InitialData/", extensions = c("csv", "xlsx"))
}

}
